Updated 11/6
Converted the rest of the nutrition categories into french with the correct accents. I also had to change some of the formatting of the program to best fit the new changes. In addition, I added the source where I found the code that helped me print out accents in python using unicode.