Created 11/10
I created this HTML nutrition label to print to our text file when we print a recipe for a user. My previous version was run in python and could print the label in the terminal but the formatting was all messed up when trying to print it to a file. I am hoping this version has more success. I also changed formatting and translations to make my nutrition label look like a standard French nutrition label once it is printed out.