Updated 11/9
This is the file I have been working on with converting our program to French. I added the nutrition label to this program and it comes out looking ok, but I think it can be improved if printed through HTML into a text file. I had to remove the class I had been using in python because it wouldn't print to the file correctly. What I have here works, but I am trying to make the label look more presentable and professional.