Updated 11/10
Made some more formatting changes and re-added the sugar and sodium fields to the print out. Made sure that everything printed out is in French in the terminal and in the text file. I formatted the Nutrition Label a little more in case I can't get my HTML version to print out correctly.